When return value of sizeof etc are being printed, typecast them to unsigned long and use %lu, instead of %d, so it works on both 32-bit an 64-bit platforms. Patch from Frank Zago.

+++ ndiswrapper/driver/misc_funcs.c
@@ -93,7 +93,7 @@ void *wrap_kmalloc(size_t size, int flags)
 	struct wrap_alloc *alloc;
 	KIRQL irql;
 
-	TRACEENTER4("size = %u, flags = %d", size, flags);
+	TRACEENTER4("size = %lu, flags = %d", (unsigned long)size, flags);
 
 	if ((flags & GFP_ATOMIC) || irqs_disabled())
 		alloc = kmalloc(sizeof(*alloc), GFP_ATOMIC);
